# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Quest Start My Day
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: pillars
    description: Project pillar operations
  - name: milestones
    description: Pillar milestone operations
  - name: tasks
    description: Daily task operations
  - name: weekly
    description: Weekly planning operations
  - name: progress
    description: Progress log operations
  - name: dashboard
    description: Dashboard summary operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /pillars:
    get:
      operationId: listPillars
      tags: [pillars]
      summary: List all project pillars
      responses:
        "200":
          description: List of pillars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pillar"
    post:
      operationId: createPillar
      tags: [pillars]
      summary: Create a new pillar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePillarBody"
      responses:
        "201":
          description: Created pillar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pillar"

  /pillars/{id}:
    patch:
      operationId: updatePillar
      tags: [pillars]
      summary: Update a pillar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePillarBody"
      responses:
        "200":
          description: Updated pillar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pillar"

  /milestones:
    get:
      operationId: listMilestones
      tags: [milestones]
      summary: List milestones for a pillar
      parameters:
        - name: pillarId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [milestones]
      summary: Create a milestone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMilestoneBody"
      responses:
        "201":
          description: Created milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"

  /milestones/{id}:
    patch:
      operationId: updateMilestone
      tags: [milestones]
      summary: Update a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMilestoneBody"
      responses:
        "200":
          description: Updated milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
    delete:
      operationId: deleteMilestone
      tags: [milestones]
      summary: Delete a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List daily tasks
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            description: Date in YYYY-MM-DD format, defaults to today
      responses:
        "200":
          description: List of daily tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a daily task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task (status, content)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /weekly:
    get:
      operationId: listWeeklyPlans
      tags: [weekly]
      summary: List weekly plans
      parameters:
        - name: weekOf
          in: query
          required: false
          schema:
            type: string
            description: Week start date YYYY-MM-DD, defaults to current week
      responses:
        "200":
          description: List of weekly plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeeklyPlan"
    post:
      operationId: createWeeklyPlan
      tags: [weekly]
      summary: Create a weekly plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWeeklyPlanBody"
      responses:
        "201":
          description: Created weekly plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyPlan"

  /weekly/{id}:
    patch:
      operationId: updateWeeklyPlan
      tags: [weekly]
      summary: Update a weekly plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWeeklyPlanBody"
      responses:
        "200":
          description: Updated weekly plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyPlan"

  /progress:
    get:
      operationId: listProgressLogs
      tags: [progress]
      summary: List progress log entries
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: List of progress log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressLog"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get today's dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/week-summary:
    get:
      operationId: getWeekSummary
      tags: [dashboard]
      summary: Get this week's task completion summary
      responses:
        "200":
          description: Weekly summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeekSummary"

  /dashboard/reentry:
    get:
      operationId: getReentryTask
      tags: [dashboard]
      summary: Get re-entry task (last unfinished or last completed)
      responses:
        "200":
          description: Re-entry task or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReentryInfo"

  /dashboard/pillar-health:
    get:
      operationId: getPillarHealth
      tags: [dashboard]
      summary: Get per-pillar health stats for this week
      responses:
        "200":
          description: List of pillar health entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PillarHealthEntry"

components:
  schemas:
    Milestone:
      type: object
      properties:
        id:
          type: integer
        pillarId:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [planned, active, blocked, complete]
        priority:
          type: string
          nullable: true
          enum: [P1, P2, P3, P4]
        targetDate:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        nextAction:
          type: string
          nullable: true
        sortOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - pillarId
        - title
        - status
        - sortOrder
        - createdAt

    CreateMilestoneBody:
      type: object
      properties:
        pillarId:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [planned, active, blocked, complete]
        priority:
          type: string
          nullable: true
          enum: [P1, P2, P3, P4]
        targetDate:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        nextAction:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required:
        - pillarId
        - title

    UpdateMilestoneBody:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [planned, active, blocked, complete]
        priority:
          type: string
          nullable: true
          enum: [P1, P2, P3, P4]
        targetDate:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        nextAction:
          type: string
          nullable: true
        sortOrder:
          type: integer

    PillarHealthEntry:
      type: object
      properties:
        pillarId:
          type: integer
        pillarName:
          type: string
        portfolioStatus:
          type: string
          nullable: true
        tasksDoneThisWeek:
          type: integer
        tasksPushedOrPassedThisWeek:
          type: integer
        daysSinceLastMovement:
          type: integer
          nullable: true
        nudge:
          type: string
          nullable: true
        warning:
          type: string
          nullable: true
      required:
        - pillarId
        - pillarName
        - tasksDoneThisWeek
        - tasksPushedOrPassedThisWeek

    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Pillar:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        priority:
          type: string
          enum: [P1, P2, P3, P4]
        description:
          type: string
          nullable: true
        isActiveThisWeek:
          type: boolean
        color:
          type: string
          nullable: true
        createdAt:
          type: string
        portfolioStatus:
          type: string
          nullable: true
        currentStage:
          type: string
          nullable: true
        whyItMatters:
          type: string
          nullable: true
        nowFocus:
          type: string
          nullable: true
        nextFocus:
          type: string
          nullable: true
        laterFocus:
          type: string
          nullable: true
        blockers:
          type: string
          nullable: true
        lastUpdated:
          type: string
          nullable: true
      required:
        - id
        - name
        - priority
        - isActiveThisWeek
        - createdAt

    CreatePillarBody:
      type: object
      properties:
        name:
          type: string
        priority:
          type: string
          enum: [P1, P2, P3, P4]
        description:
          type: string
          nullable: true
        isActiveThisWeek:
          type: boolean
        color:
          type: string
          nullable: true
        portfolioStatus:
          type: string
          nullable: true
      required:
        - name
        - priority
        - isActiveThisWeek

    UpdatePillarBody:
      type: object
      properties:
        name:
          type: string
        priority:
          type: string
          enum: [P1, P2, P3, P4]
        description:
          type: string
          nullable: true
        isActiveThisWeek:
          type: boolean
        color:
          type: string
          nullable: true
        portfolioStatus:
          type: string
          nullable: true
        currentStage:
          type: string
          nullable: true
        whyItMatters:
          type: string
          nullable: true
        nowFocus:
          type: string
          nullable: true
        nextFocus:
          type: string
          nullable: true
        laterFocus:
          type: string
          nullable: true
        blockers:
          type: string
          nullable: true
        lastUpdated:
          type: string
          nullable: true

    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [business, creative, wellness]
        whyItMatters:
          type: string
          nullable: true
        doneLooksLike:
          type: string
          nullable: true
        suggestedNextStep:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, done, pushed, passed, blocked]
        pillarId:
          type: integer
          nullable: true
        milestoneId:
          type: integer
          nullable: true
        blockerReason:
          type: string
          nullable: true
        date:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - category
        - status
        - date
        - createdAt

    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
          enum: [business, creative, wellness]
        whyItMatters:
          type: string
          nullable: true
        doneLooksLike:
          type: string
          nullable: true
        suggestedNextStep:
          type: string
          nullable: true
        pillarId:
          type: integer
          nullable: true
        milestoneId:
          type: integer
          nullable: true
        blockerReason:
          type: string
          nullable: true
        date:
          type: string
      required:
        - title
        - category
        - date

    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
          enum: [business, creative, wellness]
        whyItMatters:
          type: string
          nullable: true
        doneLooksLike:
          type: string
          nullable: true
        suggestedNextStep:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, done, pushed, passed, blocked]
        pillarId:
          type: integer
          nullable: true
        milestoneId:
          type: integer
          nullable: true
        blockerReason:
          type: string
          nullable: true

    WeeklyPlan:
      type: object
      properties:
        id:
          type: integer
        weekOf:
          type: string
        priorities:
          type: array
          items:
            type: string
        healthFocus:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        activePillarIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
        businessFocus:
          type: string
          nullable: true
        creativeFocus:
          type: string
          nullable: true
        whatMovedForward:
          type: string
          nullable: true
        whatGotStuck:
          type: string
          nullable: true
        whatContinues:
          type: string
          nullable: true
        whatToDeprioritize:
          type: string
          nullable: true
        nextWeekFocus:
          type: string
          nullable: true
      required:
        - id
        - weekOf
        - priorities
        - activePillarIds
        - createdAt

    CreateWeeklyPlanBody:
      type: object
      properties:
        weekOf:
          type: string
        priorities:
          type: array
          items:
            type: string
        healthFocus:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        activePillarIds:
          type: array
          items:
            type: integer
        businessFocus:
          type: string
          nullable: true
        creativeFocus:
          type: string
          nullable: true
        whatMovedForward:
          type: string
          nullable: true
        whatGotStuck:
          type: string
          nullable: true
        whatContinues:
          type: string
          nullable: true
        whatToDeprioritize:
          type: string
          nullable: true
        nextWeekFocus:
          type: string
          nullable: true
      required:
        - weekOf
        - priorities
        - activePillarIds

    UpdateWeeklyPlanBody:
      type: object
      properties:
        priorities:
          type: array
          items:
            type: string
        healthFocus:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        activePillarIds:
          type: array
          items:
            type: integer
        businessFocus:
          type: string
          nullable: true
        creativeFocus:
          type: string
          nullable: true
        whatMovedForward:
          type: string
          nullable: true
        whatGotStuck:
          type: string
          nullable: true
        whatContinues:
          type: string
          nullable: true
        whatToDeprioritize:
          type: string
          nullable: true
        nextWeekFocus:
          type: string
          nullable: true

    ProgressLog:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
          nullable: true
        taskTitle:
          type: string
        category:
          type: string
        status:
          type: string
        date:
          type: string
        loggedAt:
          type: string
      required:
        - id
        - taskTitle
        - category
        - status
        - date
        - loggedAt

    DashboardSummary:
      type: object
      properties:
        todayDate:
          type: string
        totalTasksToday:
          type: integer
        doneCount:
          type: integer
        pushedCount:
          type: integer
        passedCount:
          type: integer
        blockedCount:
          type: integer
        pendingCount:
          type: integer
        activePillars:
          type: array
          items:
            $ref: "#/components/schemas/Pillar"
        weeklyPlan:
          $ref: "#/components/schemas/WeeklyPlan"
          nullable: true
        planningStreak:
          type: integer
          description: Number of consecutive weeks with a weekly plan saved
      required:
        - todayDate
        - totalTasksToday
        - doneCount
        - pushedCount
        - passedCount
        - blockedCount
        - pendingCount
        - activePillars
        - planningStreak

    WeekSummary:
      type: object
      properties:
        weekOf:
          type: string
        totalTasks:
          type: integer
        doneCount:
          type: integer
        pushedCount:
          type: integer
        passedCount:
          type: integer
        blockedCount:
          type: integer
        completionRate:
          type: number
        pillarActivity:
          type: array
          items:
            type: object
            properties:
              pillarId:
                type: integer
              pillarName:
                type: string
              taskCount:
                type: integer
              tasks:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    title:
                      type: string
                    status:
                      type: string
                    category:
                      type: string
                  required:
                    - id
                    - title
                    - status
                    - category
            required:
              - pillarId
              - pillarName
              - taskCount
              - tasks
      required:
        - weekOf
        - totalTasks
        - doneCount
        - pushedCount
        - passedCount
        - blockedCount
        - completionRate
        - pillarActivity

    ReentryInfo:
      type: object
      properties:
        type:
          type: string
          enum: [unfinished, completed, none]
        task:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            title:
              type: string
            suggestedNextStep:
              type: string
              nullable: true
            whyItMatters:
              type: string
              nullable: true
            blockerReason:
              type: string
              nullable: true
            milestoneTitle:
              type: string
              nullable: true
            status:
              type: string
            date:
              type: string
            category:
              type: string
            pillarId:
              type: integer
              nullable: true
          required:
            - id
            - title
            - status
            - date
            - category
        guidance:
          type: string
          nullable: true
      required:
        - type